+++ src/main/java/com/prueba/util/ReporteComparativo.java
@@ -29,9 +29,9 @@ public class ReporteComparativo {
 	
 	private Usuario usuarioCrea;
 	
-	private Integer inv1;
+	private Long inv1;
 	
-	private Integer inv2;
+	private Long inv2;
 	
 	public ReporteComparativo(List<ComparativoInventarioDTO> comparativo, Usuario usuario) {
 		super();
